content-inspection/requirements.txt
beautifulsoup4==4.9.3
httplib2==0.19.1
boto3==1.18.44
selenium==3.141.0
requests==2.26.0
fastapi
uvicorn
chromedriver-binary==79.0.3945.36
numpy==1.19.5
opencv-python-headless==4.5.4.58
pytest
Pillow
pymongo==3.11.4
python-dotenv==1.0.0